% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shap.R
\name{check_var_names}
\alias{check_var_names}
\title{Check predictor names for test set}
\usage{
check_var_names(var_names, x_test)
}
\arguments{
\item{var_names}{A vector of variable names}

\item{x_test}{A \code{data.frame} of predictors from the test set.}
}
\description{
Check predictor names for test set
}
