% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_models.R
\name{draw_models_initial}
\alias{draw_models_initial}
\title{Generate M logistic regression models centered around the optimal model}
\usage{
draw_models_initial(
  coef_optim,
  coef_optim_var,
  x,
  y,
  M = 800,
  u1 = 0.5,
  u2 = 20
)
}
\arguments{
\item{coef_optim}{Regression coefficients of the optimal logistic regression
(i.e., the model that minimized logistic loss).}

\item{coef_optim_var}{Variance-covariance matrix of \code{coef_optim}.}

\item{x}{A \code{data.frame} of predictors from the training set. Make
sure categorical variables are properly encoded as factors.}

\item{y}{A numeric vector of outcome, with events coded as 1.}

\item{M}{Total number of models to generate.}

\item{u1}{Lower bound of a uniform distribution.}

\item{u2}{Upper bound of a uniform distribution.}
}
\value{
Returns a \code{data.frame} with \code{M} rows.
}
\description{
Generate M logistic regression models centered around the optimal model
}
