% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_shap.R
\name{find_clusters}
\alias{find_clusters}
\title{Find clusters of columns that correspond to the same variable in one-hot coding}
\usage{
find_clusters(data)
}
\arguments{
\item{data}{A \code{data.frame} of predictors.}
}
\value{
Returns a list of vectors, where the j-th vector contains the columns
  in the one-hot coding that correspond to a variable in the original
  \code{data.frame}.
}
\description{
Find clusters of columns that correspond to the same variable in one-hot coding
}
\details{
When a \code{data.frame} of predictors is converted to one-hot
  coding (i.e., a \code{model.matrix} without the column corresponding to the
  intercept term), columns that correspond to different levels of the same
  category variable belong to the same cluster. This is used to compute SHAP
  and SAGE values for categorical variables.
}
\examples{
df <- data.frame(x1 = 1:10,
                 x2 = factor(sample(letters[1:3], size = 10, replace = TRUE)),
                 x3 = factor(sample(c("f", "m"), size = 10, replace = TRUE)))
ShapleyVIC:::find_clusters(df)
}
