% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_models.R
\name{plot_perf_metric}
\alias{plot_perf_metric}
\title{Plot performance metrics of sampled models}
\usage{
plot_perf_metric(
  perf_metric,
  eligible,
  select = NULL,
  x_range,
  plot_selected = FALSE,
  x_breaks = NULL
)
}
\arguments{
\item{perf_metric}{Numeric vector of performance metrics for all sampled
models.}

\item{eligible}{Numeric vector of the same length of \code{perf_metric},
indicating whether each sample is eligible.}

\item{select}{(Optional) Numeric vector of the same length of
\code{perf_metric}, indicating whether each sample is selected in the final
sample.}

\item{x_range}{Numeric vector indicating the range of eligible values for
performance metrics. Will be indicated by dotted vertical lines in plots.}

\item{plot_selected}{Whether performance metrics of selected models should be
plotted in a second panel. Default is no.}

\item{x_breaks}{(Optional) If selected models are to be plotted, the breaks
to use in the histogram.}
}
\value{
Histogram(s) of model performance made using \code{ggplot}.
}
\description{
Plot performance metrics of sampled models
}
